+++ magpie/ui/utils.py
@@ -85,16 +85,18 @@ def error_badrequest(func):
 
 class BaseViews(object):
     """Base methods for Magpie UI pages."""
+    MAGPIE_FIXED_GROUP_MEMBERSHIPS = []
+    MAGPIE_FIXED_GROUP_EDITS = []
 
     def __init__(self, request):
         self.request = request
         self.magpie_url = get_magpie_url(request.registry)
         self.logged_user = get_logged_user(request)
 
-        self.MAGPIE_FIXED_GROUP_MEMBERSHIPS = [
-            get_constant("MAGPIE_ANONYMOUS_GROUP", settings_container=request),
-        ]
-        """Special groups membership that cannot be edited."""
+        anonymous = get_constant("MAGPIE_ANONYMOUS_GROUP", settings_container=request)
+        admin = get_constant("MAGPIE_ADMIN_GROUP", settings_container=request)
+        self.MAGPIE_FIXED_GROUP_MEMBERSHIPS = [anonymous]   # special groups membership that cannot be edited
+        self.MAGPIE_FIXED_GROUP_EDITS = [anonymous, admin]  # special groups that cannot be edited
 
     def add_template_data(self, data=None):
         # type: (Optional[Dict[Str, Any]]) -> Dict[Str, Any]
@@ -102,6 +104,7 @@ class BaseViews(object):
         all_data = data or {}
         all_data.setdefault("MAGPIE_SUB_TITLE", "Administration")
         all_data.setdefault("MAGPIE_FIXED_GROUP_MEMBERSHIPS", self.MAGPIE_FIXED_GROUP_MEMBERSHIPS)
+        all_data.setdefault("MAGPIE_FIXED_GROUP_EDITS", self.MAGPIE_FIXED_GROUP_EDITS)
         magpie_logged_user = get_logged_user(self.request)
         if magpie_logged_user:
             all_data.update({"MAGPIE_LOGGED_USER": magpie_logged_user.user_name})